Function duguo::launch_rocket

pub fn launch_rocket() -> Result<(), Box<dyn Error>>

Starts the Rocket web server and corresponding services. Called in main.rs. Note: the Tokio version is deliberately set to 0.2.24 to match the MongoDB 1.1.1 driver. No new Tokio runtimes should be created in other functions and since they can lead to runtime panics.