Trait duguo::CacheItem

pub trait CacheItem {
    fn get_uid_ordered_values(&self) -> Vec<&str>;

    fn get_uid(&self) -> String { ... }

    fn generate_uid(ordered_items: Vec<&str>) -> String { ... }
}

An object that can be found in Redis (using a uid).

Required methods

fn get_uid_ordered_values(&self) -> Vec<&str>

From object, returns values to generate uid.

Loading content...
Provided methods

fn get_uid(&self) -> String

Gets the uid used in cache lookups.

fn generate_uid(ordered_items: Vec<&str>) -> String

Generates the uid used in cache lookups. This removes all spaces, then concatenates the elements in ordered_items.

Loading content...