Trait duguo::DatabaseItem

pub trait DatabaseItem {
    fn collection_name() -> &'static str;

    fn all_field_names() -> Vec<&'static str>;

    fn primary_key(&self) -> &str;

    fn as_document(&self) -> Document
    where
        Self: Serialize,
    { ... }

    fn as_bson(&self) -> Bson
    where
        Self: Serialize,
    { ... }

    fn try_lookup(db: &Database, query_doc: Document) -> Option<Document>
    where
        Self: Sized,
    { ... }

    fn try_insert(&self, db: &Database) -> Result<String, Box<dyn Error>>
    where
        Self: Serialize,
    { ... }

    fn try_update(
        &self, 
        db: &Database, 
        keys: Vec<&str>, 
        new_values: Vec<&str>
    ) -> Result<(), Box<dyn Error>>
    where
        Self: Serialize,
    { ... }

    fn get_values_from_query(
        db: &Database, 
        query_doc: Document, 
        fields: Vec<&str>
    ) -> Vec<String> { ... }

    fn aggregate_all_values_from_query(
        db: &Database, 
        query_doc: Document, 
        fields: Vec<&str>
    ) -> Vec<Vec<String>> { ... }
}

An object that can be found in MongoDB. All DatabaseItem object fields are stored into MongoDB as String fields.

Required methods

fn collection_name() -> &'static str

Returns the collection name in MongoDB where the objects should be stored.

fn all_field_names() -> Vec<&'static str>

Returns all valid field names for the document.

fn primary_key(&self) -> &str

Returns a generally-informative key for the document. This is not guaranteed to be unique (though generally is).

Loading content...
Provided methods

fn as_document(&self) -> Document where
Self: Serialize,

Returns object as bson::document::Document type.

fn as_bson(&self) -> Bson where
Self: Serialize,

Returns object as bson::Bson type.

fn try_lookup(db: &Database, query_doc: Document) -> Option<Document> where
Self: Sized,

Attempts to lookup a full Document based on the provided query Document.

fn try_insert(&self, db: &Database) -> Result<String, Box<dyn Error>> where
Self: Serialize,

Attempts to insert the object into MongoDB.

fn try_update(
&self,
db: &Database,
keys: Vec<&str>,
new_values: Vec<&str>
) -> Result<(), Box<dyn Error>> where
Self: Serialize,

If the current object exists in MongoDB, attempts to update the corresponding key fields with new values. Input Vec<&str> are read in-order and must be the same size.

fn get_values_from_query(
db: &Database,
query_doc: Document,
fields: Vec<&str>
) -> Vec<String>

From the first result matching the query_doc, returns the values from input fields as a Vec (with matching indices as the input fields).

fn aggregate_all_values_from_query(
db: &Database,
query_doc: Document,
fields: Vec<&str>
) -> Vec<Vec<String>>

From all documents matching the query_doc, aggregates the values from the input fields into a Vec (with matching indices as the input fields). If an input field is invalid, then the corresponding Vec will be empty.

Loading content...